requirements.txt
ffmpeg-python
requests
validators
yt-dlp
python-telegram-bot